requirements.txt
# aioamqp==0.14.0
# aioredis==1.3.1
# apns2==0.7.1
# asgi-redis==1.4.3
# asgiref
# async-timeout==3.0.1
# attrs==19.3.0
# autobahn==20.4.3
# Automat==20.2.0
# autopep8==1.4.4
# blinker==1.4
# boto3
# botocore
# certifi==2019.3.9

# channels==2.4.0
# channels-rabbitmq==1.2.1
# channels-redis==2.4.2
# chardet==3.0.4
# Click==7.0
# constantly==15.1.0
# crypto==1.4.1
# cryptography==2.8
# daphne==2.5.0
# decorator==4.4.1
# decouple==0.0.7
# defusedxml==0.6.0
# dj-database-url==0.5.0
# Django==2.2.10
# django-ajax-selects==1.8.0
# django-allauth==0.41.0
# django-appconf==1.0.4
# django-braces==1.14.0
# django-classy-tags==1.0.0
# django-cors-headers==3.4.0
# django-countries==6.1.2
# django-crispy-forms==1.8.0
# django-easy-maps==1.1.0
# django-extensions==3.0.8
# django-filter==2.3.0
# django-heroku==0.3.1
# django-ipware==2.1.0
# django-jsonfield==1.4.0
# django-model-utils==4.0.0
# django-money==1.1
# django-notification==1.3.3
# django-notifications-hq==1.6.0
# django-phone-field==1.8.1
# django-push-notifications==2.0.0
# django-rating==0.1.5
# django-redis-cache==2.1.1
# django-rest-auth==0.9.5
# django-star-ratings==0.9.1
# django-storages==1.9.1
# django-taggit==1.3.0

# django-user-accounts==3.0.0
# django-user-agents==0.4.0
# django-webpush==0.3.2
# django-widget-tweaks==1.4.5
# djangorave==0.1.1
# djangorestframework==3.10.3
# dnspython==1.16.0
# docutils==0.15.2

# Flutterwave==1.0.7
# future==0.18.2
# geographiclib==1.50
# geoip2==3.0.0
# geopy==1.22.0
# gunicorn==20.0.4
# h2==2.6.2
# hiredis==1.0.1
# hpack==3.0.0
# http-ece==1.1.0
# hyper==0.7.0
# hyperframe==3.2.0
# hyperlink==19.0.0
# idna==2.8
# imageio==2.6.1
# imageio-ffmpeg==0.3.0
# importlib-metadata==1.7.0
# incremental==17.5.0
# itsdangerous==1.1.0
# Jinja2==2.11.1
# jmespath==0.10.0
# jsonfield==3.1.0
# Markdown==3.2.2
# MarkupSafe==1.1.1
# maxminddb==1.5.4
# moviepy==1.0.1
# msgpack==0.6.2
# msgpack-python==0.5.6
# mysql-connector-python==8.0.19
# mysqlclient==1.4.5
# Naked==0.1.31
# numpy==1.18.1
# oauthlib==3.1.0
# opencv-contrib-python==4.4.0.42
# opencv-python==4.4.0.42
# pamqp==2.3.0
# passlib==1.7.2
# Pillow==7.0.0
# proglog==0.1.9
# protobuf==3.6.1
# psycopg2==2.8.5
# py-moneyed==0.8.0
# py-vapid==1.7.0
# pyasn1==0.4.8
# pyasn1-modules==0.2.8
# pycodestyle==2.5.0
# pycparser==2.20
# pycrypto==2.6.1
# pygeoip==0.3.2
# PyHamcrest==2.0.2
# PyJWT==1.7.1
# PyMySQL==0.9.3
# pyOpenSSL==19.1.0
# python-dateutil==2.8.1
# python-decouple==3.3
# python-resize-image==1.1.19
# python3-openid==3.1.0
# pytz==2018.5
# pywebpush==1.9.4
# PyYAML==5.3.1
# redis==2.10.6
# requests==2.22.0
# requests-oauthlib==1.3.0
# s3transfer==0.3.3
# service-identity==18.1.0
# shellescape==3.8.1
# six==1.13.0
# sorl-thumbnail==12.6.3
# SQLAlchemy==1.3.0
# sqlparse==0.2.4
# swapper==1.1.2.post1
# tqdm==4.41.1
# Twisted==20.3.0
# txaio==20.4.1
# ua-parser==0.10.0
# urllib3==1.25.7
# user-agents==2.1
# Werkzeug==0.16.0
# whitenoise==5.0.1
# WTForms==2.2.1
# zipp==3.1.0
# zope.interface==5.1.0


boto3==1.4.6
botocore==1.6.8
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
coverage==4.5.3
cryptography==36.0.1
defusedxml==0.7.1
Django==2.2.13
django-allauth==0.47.0
django-ckeditor==6.0.0
django-crispy-forms==1.13.0
django-debug-toolbar==1.11
django-filter==2.1.0
django-js-asset==1.2.2
django-model-utils==4.2.0
django-storages==1.11.1
django-star-ratings==0.9.1
djangorave==0.1.1
django-taggit==1.1.0
django-webpush==0.3.4
djangorestframework==3.9.2
django-extensions==3.0.8
django-classy-tags==1.0.0
django-cors-headers==3.4.0
docutils==0.18.1
gunicorn==20.1.0
http-ece==1.1.0
idna==3.3
jmespath==0.10.0
jsonfield==3.1.0
model-mommy==1.6.0
mysqlclient==2.1.0
oauthlib==3.1.1
Pillow==8.4.0
py-vapid==1.8.2
pycparser==2.21
django-phone-field==1.8.1
django-push-notifications==2.0.0
django-rating==0.1.5
django-rest-auth==0.9.5
PyJWT==2.3.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
pywebpush==1.9.4
requests==2.27.1
# Flutterwave==1.0.7
requests-oauthlib==1.3.0
s3transfer==0.1.13
six==1.16.0
sqlparse==0.4.2
swapper==1.3.0
urllib3==1.26.7
whitenoise==5.3.0
